# swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "SeqUI Server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/seqapi/v1/async_search/fetch": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_fetch_async_search_result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.FetchAsyncSearchResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.FetchAsyncSearchResultResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/seqapi/v1/export": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.ExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful streaming responses",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.ExportResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/seqapi/v1/fields": {
            "get": {
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_getFields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetFieldsResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/seqapi/v1/fields/pinned": {
            "get": {
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_getPinnedFields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetFieldsResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "UnexpectedError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "http.AsyncSearchRequestHistogram": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "string"
                }
            }
        },
...
        "seqapi.v1.Aggregation": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationBucket"
                    }
                },
                "not_exists": {
                    "type": "integer"
                }
            }
        },
        "seqapi.v1.AggregationBucket": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "not_exists": {
                    "type": "integer"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "seqapi.v1.AggregationFunc": {
            "type": "string",
            "enum": [
                "count",
                "sum",
                "min",
                "max",
                "avg",
                "quantile",
                "unique"
            ],
            "x-enum-varnames": [
                "afCount",
                "afSum",
                "afMin",
                "afMax",
                "afAvg",
                "afQuantile",
                "afUnique"
            ]
        },
...
        "seqapi.v1.AggregationSeries": {
            "type": "object",
            "properties": {
                "metric": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationTsBucket"
                    }
                }
            }
        },
        "seqapi.v1.AggregationTs": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "result": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/seqapi.v1.AggregationSeries"
                            }
                        }
                    }
                }
            }
        },
        "seqapi.v1.AggregationTsBucket": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "integer"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "seqapi.v1.AggregationTsQuery": {
            "type": "object",
            "properties": {
                "agg_func": {
                    "default": "count",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.AggregationFunc"
                        }
                    ]
                },
                "field": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "interval": {
                    "type": "string",
                    "format": "duration",
                    "example": "1m"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "seqapi.v1.AsyncSearchResponse": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Aggregation"
                    }
                },
                "aggregations_ts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationTs"
                    }
                },
                "error": {
                    "$ref": "#/definitions/seqapi.v1.Error"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Event"
                    }
                },
                "histogram": {
                    "$ref": "#/definitions/seqapi.v1.Histogram"
                },
                "partialResponse": {
                    "type": "boolean"
                },
                "total": {
                    "type": "string",
                    "format": "int64"
                }
            }
        },
        "seqapi.v1.AsyncSearchStatus": {
            "type": "string",
            "enum": [
                "in_progress",
                "done",
                "canceled",
                "error"
            ],
            "x-enum-varnames": [
                "AsyncSearchStatusInProgress",
                "AsyncSearchStatusDone",
                "AsyncSearchStatusCanceled",
                "AsyncSearchStatusError"
            ]
        },
...
        "seqapi.v1.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "default": "ERROR_CODE_NO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.ErrorCode"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "seqapi.v1.ErrorCode": {
            "type": "string",
            "enum": [
                "ERROR_CODE_NO",
                "ERROR_CODE_PARTIAL_RESPONSE",
                "ERROR_CODE_QUERY_TOO_HEAVY",
                "ERROR_CODE_TOO_MANY_FRACTIONS_HIT"
            ],
            "x-enum-varnames": [
                "aecNo",
                "aecPartialResponse",
                "aecQueryTooHeavy",
                "aecTooManyFractionsHit"
            ]
        },
        "seqapi.v1.Event": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "time": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "seqapi.v1.ExportFormat": {
            "type": "string",
            "enum": [
                "jsonl",
                "csv"
            ],
            "x-enum-varnames": [
                "efJSONL",
                "efCSV"
            ]
        },
        "seqapi.v1.ExportRequest": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "format": {
                    "default": "jsonl",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.ExportFormat"
                        }
                    ]
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "limit": {
                    "type": "integer",
                    "format": "int32"
                },
                "offset": {
                    "type": "integer",
                    "format": "int32"
                },
                "query": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "seqapi.v1.ExportResponse": {
            "description": "Export response in one of the following formats:\u003cbr\u003e - JSONL: {\"id\":\"some-id\",\"data\":{\"field1\":\"value1\",\"field2\":\"value2\"},\"time\":\"2024-12-31T10:20:30.0004Z\"}\u003cbr\u003e - CSV: value1,value2,value3",
            "type": "object"
        },
        "seqapi.v1.FetchAsyncSearchResultRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "format": "int32"
                },
                "offset": {
                    "type": "integer",
                    "format": "int32"
                },
                "order": {
                    "default": "desc",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.Order"
                        }
                    ]
                },
                "search_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "seqapi.v1.FetchAsyncSearchResultResponse": {
            "type": "object",
            "properties": {
                "canceled_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "disk_usage": {
                    "type": "string",
                    "format": "int64"
                },
                "expires_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "meta": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                },
                "request": {
                    "$ref": "#/definitions/seqapi.v1.StartAsyncSearchRequest"
                },
                "response": {
                    "$ref": "#/definitions/seqapi.v1.AsyncSearchResponse"
                },
                "started_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "$ref": "#/definitions/seqapi.v1.AsyncSearchStatus"
                }
            }
        },
        "seqapi.v1.Field": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "default": "unknown",
                    "enum": [
                        "unknown",
                        "keyword",
                        "text"
                    ]
                }
            }
        },
...
        "seqapi.v1.GetFieldsResponse": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Field"
                    }
                }
            }
        },
...
        "seqapi.v1.Histogram": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.HistogramBucket"
                    }
                }
            }
        },
        "seqapi.v1.HistogramBucket": {
            "type": "object",
            "properties": {
                "docCount": {
                    "type": "string",
                    "format": "uint64"
                },
                "key": {
                    "type": "string",
                    "format": "uint64"
                }
            }
        },
        "seqapi.v1.Order": {
            "type": "string",
            "enum": [
                "desc",
                "asc"
            ],
            "x-enum-varnames": [
                "oDESC",
                "oASC"
            ]
        },
...
        "seqapi.v1.StartAsyncSearchRequest": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationTsQuery"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "histogram": {
                    "$ref": "#/definitions/http.AsyncSearchRequestHistogram"
                },
                "meta": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "retention": {
                    "type": "string",
                    "format": "duration",
                    "example": "1h"
                },
                "size": {
                    "type": "integer"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "with_docs": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "description": "Authentication token, prefixed by Bearer: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}